backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
numpy>=1.26.0
scipy>=1.12.0
librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1
moviepy>=2.0.0
ffmpeg-python>=0.2.0
Pillow>=10.2.0
gunicorn>=21.2.0
pytest>=8.0.0
flake8>=7.0.0
